+++ src/index.ts
@@ -121,6 +121,7 @@ export class PropulsionAI extends Core.APIClient {
 
   chat: API.Chat = new API.Chat(this);
   dataset: API.Dataset = new API.Dataset(this);
+  knowledgebase: API.Knowledgebase = new API.Knowledgebase(this);
 
   protected override defaultQuery(): Core.DefaultQuery | undefined {
     return this._options.defaultQuery;
@@ -184,6 +185,10 @@ export namespace PropulsionAI {
   export import Dataset = API.Dataset;
   export import DatasetCreateResponse = API.DatasetCreateResponse;
   export import DatasetCreateParams = API.DatasetCreateParams;
+
+  export import Knowledgebase = API.Knowledgebase;
+  export import KnowledgebaseCreateResponse = API.KnowledgebaseCreateResponse;
+  export import KnowledgebaseCreateParams = API.KnowledgebaseCreateParams;
 }
 
 export default PropulsionAI;