+++ src/resources/index.ts
@@ -1,3 +1,8 @@
 export { Chat } from './chat/chat';
 export { CompletionCreateResponse, CompletionCreateParams, Completions } from './chat/completions';
 export { DatasetCreateResponse, DatasetCreateParams, Dataset } from './dataset/dataset';
+export {
+  KnowledgebaseCreateResponse,
+  KnowledgebaseCreateParams,
+  Knowledgebase,
+} from './knowledgebase/knowledgebase';