+++ src/resources/knowledgebase/index.ts
@@ -1,0 +1,4 @@
+// File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
+
+export { FileDeleteResponse, FileUploadResponse, FileUploadParams, File } from './file';
+export { KnowledgebaseCreateResponse, KnowledgebaseCreateParams, Knowledgebase } from './knowledgebase';